COMMERCE-9619 Move helpers to the right package
